chore: use ruff for linting; add precommit
