--- .flake8
@@ -1,5 +1,0 @@
-[flake8]
-ignore = E203, F821
-exclude = .git,__pycache__,docs/source/conf.py,old,build,dist,proto
-max-complexity = 10
-max-line-length = 100
