+++ .github/workflows/ci.yml
@@ -34,31 +34,3 @@ jobs:
 
     - name: Run tests
       run: make test
-
-  format:
-    name: Black format
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.9"]
-
-    steps:
-    - uses: actions/checkout@v3
-
-    - name: Install poetry
-      run: pipx install poetry
-
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-        cache: 'poetry'
-
-    - name: Install dependencies
-      run: |
-        poetry env use ${{ matrix.python-version }}
-        poetry install --with dev --no-root
-
-    - name: Black format check
-      run: poetry run black --check .
