+++ .gitignore
@@ -169,3 +169,6 @@ protos/
 
 # Generated by IDE
 *.iml
+
+# Generated by ruff
+.ruff_cache
