+++ examples/function/asyncio-reduce/async_exec.py
@@ -3,7 +3,7 @@ import aiorun
 import asyncio
 import json
 import time
-from typing import AsyncIterable, List
+from collections.abc import AsyncIterable
 
 from pynumaflow import setup_logging
 from pynumaflow.function import (
@@ -28,7 +28,7 @@ async def http_request(session, url):
             return "Error"
 
 
-async def reduce_handler(keys: List[str], datums: AsyncIterable[Datum], md: Metadata) -> Messages:
+async def reduce_handler(keys: list[str], datums: AsyncIterable[Datum], md: Metadata) -> Messages:
     interval_window = md.interval_window
     async with aiohttp.ClientSession() as session:
         tasks = []
