+++ examples/function/multiproc_map/example.py
@@ -1,5 +1,4 @@
 import math
-from typing import List
 
 from pynumaflow.function import Messages, Message, Datum, MultiProcServer
 
@@ -12,7 +11,7 @@ def is_prime(n):
         return True
 
 
-def my_handler(keys: List[str], datum: Datum) -> Messages:
+def my_handler(keys: list[str], datum: Datum) -> Messages:
     val = datum.value
     _ = datum.event_time
     _ = datum.watermark
