+++ examples/sink/async_log/example.py
@@ -1,4 +1,4 @@
-from typing import AsyncIterable
+from collections.abc import AsyncIterable
 
 import aiorun
