+++ pynumaflow/function/multiproc_server.py
@@ -4,7 +4,8 @@ import multiprocessing
 import os
 import socket
 from concurrent import futures
-from typing import Callable, AsyncIterable, List
+from typing import Callable
+from collections.abc import AsyncIterable
 
 import grpc
 from google.protobuf import empty_pb2 as _empty_pb2
@@ -38,9 +39,9 @@ _LOGGER = setup_logging(__name__)
 if os.getenv("PYTHONDEBUG"):
     _LOGGER.setLevel(logging.DEBUG)
 
-UDFMapCallable = Callable[[List[str], Datum], Messages]
-UDFMapTCallable = Callable[[List[str], Datum], MessageTs]
-UDFReduceCallable = Callable[[List[str], AsyncIterable[Datum], Metadata], Messages]
+UDFMapCallable = Callable[[list[str], Datum], Messages]
+UDFMapTCallable = Callable[[list[str], Datum], MessageTs]
+UDFReduceCallable = Callable[[list[str], AsyncIterable[Datum], Metadata], Messages]
 
 
 class MultiProcServer(udfunction_pb2_grpc.UserDefinedFunctionServicer):
