+++ pyproject.toml
@@ -26,6 +26,7 @@ grpcio-tools = "^1.48.1"
 google-cloud = "^0.34.0"
 google-api-core = "^2.11.0"
 protobuf = ">=3.20,<5.0"
+aiorun = "^2022.11.1"
 
 [tool.poetry.group.dev]
 optional = true
@@ -33,9 +34,11 @@ optional = true
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2.1"
 pytest-cov = "^3.0"
-black = "^22.1"
+black = "^23.1"
 grpcio-testing = "^1.48.1"
 flake8 = "^4.0.1"
+ruff = "^0.0.264"
+pre-commit = "^3.3.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
@@ -43,6 +46,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.black]
 line-length = 100
+target-version = ["py39"]
 include = '\.pyi?$'
 exclude = '''
 /(
@@ -58,5 +62,20 @@ exclude = '''
   | build
   | dist
   | tests/.*/setup.py
+  | function/proto
+  | sink/proto
 )/
 '''
+
+[tool.ruff]
+line-length = 100
+src = ["pynumaflow", "tests", "examples"]
+select = ["E", "F", "UP"]
+target-version = "py39"
+show-fixes = true
+show-source = true
+extend-exclude = [
+    "*_pb2.py",
+    "*_pb2*.py",
+    "*.pyi"
+]
