+++ tests/sink/test_server.py
@@ -1,6 +1,6 @@
 import unittest
 from datetime import datetime, timezone
-from typing import Iterator
+from collections.abc import Iterator
 
 from google.protobuf import empty_pb2 as _empty_pb2
 from google.protobuf import timestamp_pb2 as _timestamp_pb2
